requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
Requests==2.32.3
SQLAlchemy==2.0.32
SQLAlchemy==2.0.34
